init(file-manager): finish function in the up.js, cd.js, ls.js
