+++ src/baseSwitch/index.js
@@ -1,0 +1,1 @@
+export { baseSwitch } from "./baseSwitch.js";
