+++ src/commands/index.js
@@ -1,5 +1,4 @@
 export { startProgram } from "./startProgram.js";
 export { getUsername } from "./getUsername.js";
 export { getPathToWorkDirectory } from "./getPathToWorkDirectory.js";
-export { runManager } from "./runManager.js";
 export { endProgram } from "./endProgram.js";
