+++ src/navigation/cd.js
@@ -4,13 +4,17 @@ import { chdir, cwd } from "process";
 import { getPathToWorkDirectory } from "../commands/index.js";
 
 export const cd = (pathDirectory) => {
-  if (path.isAbsolute(pathDirectory)) {
-    chdir(path.join(pathDirectory));
+  try {
+    if (path.isAbsolute(pathDirectory)) {
+      chdir(path.join(pathDirectory));
 
-    console.log(getPathToWorkDirectory(pathDirectory));
-  } else {
-    chdir(path.join(cwd(), pathDirectory));
+      console.log(getPathToWorkDirectory(pathDirectory));
+    } else {
+      chdir(path.join(cwd(), pathDirectory));
 
-    console.log(getPathToWorkDirectory(cwd()));
+      console.log(getPathToWorkDirectory(cwd()));
+    }
+  } catch {
+    console.log("Operation failed");
   }
 };
